[update] better error handling
* src/main.rs: better error handling
    - check if name is empty or not
    - check if age is empty or not
    - check if age is a number or not
